+++ Libraries/PID/PID.h
@@ -13,7 +13,7 @@ class PID
   #define REVERSE  1
 
   //commonly used functions **************************************************************************
-    PID(volatile int32_t*, volatile int32_t*, volatile int32_t*,        // * constructor.  links the PID to the Input, Output, and 
+    PID(volatile int16_t*, volatile int16_t*, volatile int16_t*,        // * constructor.  links the PID to the Input, Output, and 
         double, double, double, int);     //   Setpoint.  Initial tuning parameters are also set here
 	
     void SetMode(int Mode);               // * sets PID to either Manual (0) or Auto (non-0)
@@ -63,9 +63,9 @@ class PID
 
 	int controllerDirection;
 
-    volatile int32_t *myInput;              // * Pointers to the Input, Output, and Setpoint variables
-    volatile int32_t *myOutput;             //   This creates a hard link between the variables and the 
-    volatile int32_t *mySetpoint;           //   PID, freeing the user from having to constantly tell us
+    volatile int16_t *myInput;              // * Pointers to the Input, Output, and Setpoint variables
+    volatile int16_t *myOutput;             //   This creates a hard link between the variables and the 
+    volatile int16_t *mySetpoint;           //   PID, freeing the user from having to constantly tell us
                                   //   what these values are.  with pointers we'll just know.
 			  
 	unsigned long lastTime;
